requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
openai==1.3.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
geopy==2.3.0
folium==0.14.0
beautifulsoup4==4.12.2
nltk==3.8.1
textblob==0.17.1
plotly==5.15.0
dash==2.11.1
gunicorn==21.2.0
